requirements.txt
streamlit
replicate
requests
Pillow
